refactor: WriteDoneButtonLayout으로 작성 로직 분리
