+++ fourtoon-cookie/src/pages/DiaryWritePage/DiaryWritePage.tsx
@@ -1,4 +1,4 @@
-import { KeyboardAvoidingView, Platform, SafeAreaView, View } from "react-native";
+import { SafeAreaView, View } from "react-native";
 import { useEffect, useState } from "react";
 import { NativeStackScreenProps } from "@react-navigation/native-stack";
 
@@ -12,34 +12,27 @@ import { GlobalErrorInfoType } from "../../types/error";
 import { LocalDate } from "@js-joda/core";
 
 import { RuntimeError } from "../../error/RuntimeError";
-import Button from "../../components/common/Button/Button";
-import { OS } from "../../types/os";
 import handleError from "../../error/errorhandler";
-import { ApiError } from "../../error/ApiError";
-import { API_STATUS } from "../../constants/api";
 import { useSelectedCharacterStore } from "../../store/selectedCharacter";
 import { useDiaryListStore } from "../../store/diaryList";
 import { Diary } from "../../types/diary";
+import WriteDoneButtonLayout from "./WriteDoneButtonLayout/WriteDoneButtonLayout";
+import { NavigationProp, useNavigation } from "@react-navigation/native";
 
 
 export type DiaryWritePageProp = NativeStackScreenProps<RootStackParamList, 'DiaryWritePage'>;
 
-const DiaryWritePage = ({ navigation, route }: DiaryWritePageProp) => {
+const DiaryWritePage = ({ route }: DiaryWritePageProp) => {
     const { currentDiaryId, ...rest } = route.params;
 
+    const navigation = useNavigation<NavigationProp<RootStackParamList>>();
     const { getDiaryById } = useDiaryListStore();
+    const { selectedCharacter } = useSelectedCharacterStore();
 
     const currentDiary: Diary | undefined = currentDiaryId ? getDiaryById(currentDiaryId) : undefined
     
     const [diaryDate, setDiaryDate] = useState<LocalDate>(currentDiary ? currentDiary.diaryDate : LocalDate.now());
     const [content, setContent] = useState<string>(currentDiary ? currentDiary.content : "");
-    const [isWorking, setIsWorking] = useState<boolean>(false);
-
-    const { selectedCharacter } = useSelectedCharacterStore();
-
-    const { postDiary, updateDiary } = useDiaryListStore();
-
-    const isNextButtonEnabled: boolean = content.length > 0;
     
     useEffect(() => {
         if (! selectedCharacter) {
@@ -53,58 +46,13 @@ const DiaryWritePage = ({ navigation, route }: DiaryWritePageProp) => {
         }
     }, [selectedCharacter]);
 
-    if (! selectedCharacter) {
-        return null;
-    }
-
     const handleDiaryDateChange = (newDate: LocalDate) => {
         setDiaryDate(newDate);
     }
 
-    const handleWriteDoneButtonPress = async () => {
-        if (isWorking) return;
-
-        setIsWorking(true);
-
-        const diary: Diary = currentDiary ? currentDiary : {
-            diaryId: -1,
-            content: content,
-            isFavorite: false,
-            diaryDate: diaryDate,
-            paintingImageUrls: [],
-            characterId: selectedCharacter.id
-        }
-
-        try {
-            if (! currentDiaryId) {
-                await postDiary(diary);
-            } else {
-                await updateDiary(diary);
-            }
-
-            navigation.navigate('DiaryTimelinePage');
-        } catch (error) {
-            if (error instanceof ApiError && error.getStatus() === API_STATUS.CONFLICT) {
-                error.message = "선택한 날짜에 이미 일기가 존재합니다. 다른 날을 선택해주세요.";
-            }
-
-            if (error instanceof Error) {
-                handleError(
-                    error,
-                    GlobalErrorInfoType.ALERT
-                );
-            }
-        } finally {
-            setIsWorking(false);
-        }
-    }
-
     const handleInputTextChange = (text: string) => {
-        if (isWorking) return;
-
         setContent(text);
     }
-
 
     return (
         <SafeAreaView style={S.styles.safeArea}>
@@ -119,22 +67,11 @@ const DiaryWritePage = ({ navigation, route }: DiaryWritePageProp) => {
                     onTextChange={handleInputTextChange}
                 /> 
                 <View style={S.styles.separator} />
-                <KeyboardAvoidingView 
-                    style={S.styles.bottomContainer} 
-                    enabled={true}
-                    keyboardVerticalOffset={80}
-                    behavior={(Platform.OS == OS.IOS) ? 'padding' : 'height'}
-                >
-                    <Button
-                        title="다음"
-                        onPress={handleWriteDoneButtonPress}
-                        style={{
-                            ...S.styles.nextButton, 
-                            backgroundColor: isNextButtonEnabled ? '#FFC426' : '#DDDDDD'
-                        }}
-                        textStyle={S.styles.nextButtonText}
-                    />
-                </KeyboardAvoidingView>
+                <WriteDoneButtonLayout
+                    diaryDate={diaryDate}
+                    content={content}
+                    currentDiaryId={currentDiaryId}
+                />
             </View>
         </SafeAreaView>
     );
