+++ fourtoon-cookie/src/pages/DiaryWritePage/WriteDoneButtonLayout/WriteDoneButtonLayout.styled.ts
@@ -1,0 +1,21 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+    bottomContainer: {
+        position: 'absolute',
+        bottom: 20,
+        left: 20,
+        right: 20,
+      },
+    nextButton: {
+        width: '100%',
+        height: 60,
+        borderRadius: 16,
+        justifyContent: 'center',
+        alignItems: 'center',
+      },
+      nextButtonText: {
+        fontSize: 17,
+        fontWeight: '600'
+      }
+});
