+++ fourtoon-cookie/src/pages/DiaryWritePage/WriteDoneButtonLayout/WriteDoneButtonLayout.tsx
@@ -1,0 +1,96 @@
+import { KeyboardAvoidingView, Platform } from "react-native"
+import { LocalDate } from "@js-joda/core";
+import { useState } from "react";
+import { NavigationProp, useNavigation } from "@react-navigation/native";
+import { OS } from "../../../types/os"
+import { Diary } from "../../../types/diary";
+import { useSelectedCharacterStore } from "../../../store/selectedCharacter";
+import { useDiaryListStore } from "../../../store/diaryList";
+import { RootStackParamList } from "../../../constants/routing";
+import { API_STATUS } from "../../../constants/api";
+import { ApiError } from "../../../error/ApiError";
+import handleError from "../../../error/errorhandler";
+import { GlobalErrorInfoType } from "../../../types/error";
+import Button from "../../../components/common/Button/Button";
+
+import * as S from "./WriteDoneButtonLayout.styled";
+
+export interface WriteDoneButtonLayout {
+    diaryDate: LocalDate;
+    content: string;
+    currentDiaryId?: number;
+}
+
+const WriteDoneButtonLayout = (props: WriteDoneButtonLayout) => {
+    const { diaryDate, content, currentDiaryId, ...rest } = props;
+
+    const navigation = useNavigation<NavigationProp<RootStackParamList>>();
+
+    const [isWorking, setIsWorking] = useState<boolean>(false);
+
+    const { postDiary, updateDiary } = useDiaryListStore();
+    const { selectedCharacter } = useSelectedCharacterStore();
+
+    const isNextButtonEnabled = content.length > 0 && ! isWorking;
+
+    if (! selectedCharacter) return null;
+
+    const handleWriteDoneButtonPress = async () => {
+        if (isWorking) return;
+
+        setIsWorking(true);
+
+        const diary: Diary = {
+            diaryId: currentDiaryId ? currentDiaryId : -1,
+            content: content,
+            isFavorite: false,
+            diaryDate: diaryDate,
+            paintingImageUrls: [],
+            characterId: selectedCharacter.id
+        }
+
+        try {
+            if (! currentDiaryId) {
+                await postDiary(diary);
+            } else {
+                await updateDiary(diary);
+            }
+
+            navigation.navigate('DiaryTimelinePage');
+        } catch (error) {
+            if (error instanceof ApiError && error.getStatus() === API_STATUS.CONFLICT) {
+                error.message = "선택한 날짜에 이미 일기가 존재합니다. 다른 날을 선택해주세요.";
+            }
+
+            if (error instanceof Error) {
+                handleError(
+                    error,
+                    GlobalErrorInfoType.ALERT
+                );
+            }
+        } finally {
+            setIsWorking(false);
+        }
+    }
+
+    return (
+        <KeyboardAvoidingView 
+            style={S.styles.bottomContainer} 
+            enabled={true}
+            keyboardVerticalOffset={80}
+            behavior={(Platform.OS == OS.IOS) ? 'padding' : 'height'}
+        >
+            <Button
+                title="완료"
+                onPress={handleWriteDoneButtonPress}
+                style={{
+                    ...S.styles.nextButton, 
+                    backgroundColor: isNextButtonEnabled ? '#FFC426' : '#DDDDDD'
+                }}
+                textStyle={S.styles.nextButtonText}
+            />
+        </KeyboardAvoidingView>
+    )
+}
+
+export default WriteDoneButtonLayout;
